Move tls_collect_extensions() into a separate file
Subsequent commits will pull other extensions code into this file.
